ユーザーID取得をhooksにした
