+++ client/src/pages/card.tsx
@@ -1,8 +1,7 @@
 import type { NextPage } from 'next'
-import { parseCookies, setCookie } from 'nookies'
 import { useEffect, useState } from 'react'
 import io from 'socket.io-client'
-import { getUniqueStr } from 'src/libs/common'
+import { useUserID } from 'src/hooks/useUserID'
 
 type CardNum = 0 | 1 | 2 | 3 | 5 | 8 | 13 | 21 | 44
 type CardType = {
@@ -14,37 +13,10 @@ const URI = 'http://localhost:3001'
 const socket = io(URI)
 
 const Home: NextPage = () => {
-  const [isConnected, setIsConnected] = useState(false)
   const cardsNum: CardNum[] = [0, 1, 2, 3, 5, 8, 13, 21, 44]
+  const userID = useUserID()
   const [users, setUsers] = useState<CardType[]>([])
-  const [userID, setUserID] = useState<string | null>(null)
-
-  const getUserIDFromCookies = () => {
-    const cookies = parseCookies()
-    const _userID = cookies.userID as string | undefined
-    return _userID || null
-  }
-
-  const setUserIDToCookies = (value: string) => {
-    setCookie(null, 'userID', value, {
-      maxAge: 30 * 24 * 60 * 60,
-      path: '/',
-    })
-  }
-
-  useEffect(() => {
-    if (typeof userID === 'string') return
-
-    const _userID = getUserIDFromCookies()
-    if (typeof _userID === 'string') {
-      setUserID(_userID)
-      return
-    }
-
-    const newUserID = getUniqueStr()
-    setUserID(newUserID)
-    setUserIDToCookies(newUserID)
-  }, [userID])
+  const [isConnected, setIsConnected] = useState(false)
 
   useEffect(() => {
     socket.on('connect', () => {
@@ -73,8 +45,6 @@ const Home: NextPage = () => {
 
   const cardClick = (event: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
     event.preventDefault()
-    if (userID === null) return
-
     const cardNumStr = event.currentTarget.dataset.num
     if (!cardNumStr) return
     const cardNum = parseInt(cardNumStr) as CardNum
